// hgmall-frontend/src/api/generated-admin/models/merchant-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall Admin API 文档
 * TheMall 商城管理系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ShopSimpleVO } from './shop-simple-vo';

export interface MerchantVO {
    'merchantId'?: number;
    'merchantName'?: string;
    'owner'?: string;
    'userId'?: number;
    'userName'?: string;
    'userEmail'?: string;
    'userPhone'?: string;
    'status'?: number;
    'statusText'?: string;
    'shopCount'?: number;
    'shops'?: Array<ShopSimpleVO>;
    'createTime'?: string;
    'updateTime'?: string;
}