// hgmall-frontend/src/api/generated-admin/models/user-list-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall Admin API 文档
 * TheMall 商城管理系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserVO } from './user-vo';

export interface UserListVO {
    'users'?: Array<UserVO>;
    'total'?: number;
    'page'?: number;
    'size'?: number;
}