// hgmall-frontend/src/api/generated-admin/models/user-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall Admin API 文档
 * TheMall 商城管理系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UserVO {
    'userId'?: number;
    'userName'?: string;
    'email'?: string;
    'phoneNumber'?: string;
    'roleId'?: number;
    'roleCode'?: string;
    'roleName'?: string;
    'status'?: number;
    'statusText'?: string;
    'merchantId'?: number;
    'merchantName'?: string;
    'orderCount'?: number;
    'createTime'?: string;
    'updateTime'?: string;
}