// hgmall-frontend/src/api/generated-admin/models/merchant-order-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall Admin API 文档
 * TheMall 商城管理系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AddressVO } from './address-vo';
// May contain unused imports in some cases
// @ts-ignore
import type { MerchantOrderItemVO } from './merchant-order-item-vo';

export interface MerchantOrderVO {
    'orderId'?: number;
    'batchOrderId'?: number;
    'userId'?: number;
    'userName'?: string;
    'userEmail'?: string;
    'shopId'?: number;
    'shopName'?: string;
    'status'?: number;
    'statusText'?: string;
    'total'?: number;
    'addressId'?: number;
    'address'?: AddressVO;
    'orderItems'?: Array<MerchantOrderItemVO>;
    'createTime'?: string;
    'updateTime'?: string;
}