// hgmall-frontend/src/api/generated-admin/models/audit-log-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall Admin API 文档
 * TheMall 商城管理系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface AuditLogVO {
    'auditId'?: number;
    'adminUserId'?: number;
    'adminUserName'?: string;
    'targetId'?: number;
    'targetType'?: string;
    'operationType'?: string;
    'operationDesc'?: string;
    'createTime'?: string;
}